// src/app/models/did.ts
export class DID {
  "id"?: number;
  "directInwardNumber": string;
  "directInwardDialingType": string;
  "country": string;
  "state": string;
  "city": string;
  "zip": string;
  "countryCode": string;
  "countryShortCode": string;
  "premium": boolean;
  "targetType": string;
  "forwardToNumber": string;
  "deviceType": string;
  "destinationDevice": string;
  "active": boolean;
}